Moved cancellation detection after the iteration call to avoid a possible defect with cancellation before first iteration.
Added default runaway limit and an ability to override it with a constructor
argument. Added suppression of runaway detection if it's zero.

Removed completion_notifier from the Long_Task class and made it an argument to
the task generator.

Changed counter from local variable to class member in anticipation of using
setTimeout() to allow other event handlers to run.

New documentation.