Set a high default runaway limit.
Converted from single loop to timer-spawned loop. Created '_run_once()' out of
the main 'while' loop from 'run()'. Kept the 'while' structure so as not to
rewrite all the 'break' statements at the outset. Loop only runs once because
there's a timer call and then 'return' at the end of the loop body.

Got rid of redundancy in the log messages of the counter task.