Added new UI elements. Separate source and reporting possibilities.
Started implementing a new top-level loop construct using a single iterator for
instructions. Added instruction.js to provide a base class object.

Removed some debugging statements from dealing with require() problems.

Added overall try{} block for processNode. Since this is a shim into ABP, it
should be transparent in the case of unexpected exceptions.

Changed logger to deliver a warning message (not an error) and report the
location of the caller (not the log function itself).