Removed extraneous logging message in the Crawl_Display log function (a user- visible log, not a debugging one).
Beginning of progress messages for the crawler in its log window. It prints the
instruction in JSON form. Added try-catch statement around the crawler task so
that cleanup can be done sanely.

Added a Default_Instruction constructor and use it as the yield type for the
basic instruction iterator. Renamed member "browse" to "target". Changed member
toJSON() functions to to_JSON_value(), since that's more descriptive of how the
recursive construction works. Note that the JSON value of an object is valid
JSON, but the JSON value of a string alone is not.

Minor jsdoc tweaks for some optional arguments.