Added a parallel check in the crawler to check that the tabbed_browser has an available request slot whenever the (local) 'loading' flag says it should.
Added a Set to Tabbed_Browser for allocated requests that haven't begun. This is
to ensure that we can close them if necessary.

Added a WeakMap from 'browser' objects to a collection including the Browser_Tab
itself. This map will be used to isolate the routing events and callback to the
correct objects.

Added three Tabbed_Browser methods for child tabs to call during the loading
process. They maintain the Set, the WeakMap, and the count of pending requests.