Added a new function 'unloader' function to crawler_ui.js. Renamed the old one '..._legacy'. Close the two important globals, 'current_crawler' and 'current_crawl' therein.
Changed the WeakMap in the Tabbed_Browser class to an ordinary Map. This is
because in order to close everything correctly, we need strong links to each of
the tabs.

Added close() functions to Tabbed_Browser, Browser_Tab, and Crawler.

Fixed typo "map_brower_to_child", renaming it "map_browser_to_child".

Added some logging annotations to the top level progress handler. Much of that
unpacking the flags in the progress state.

Changed 'leave_open' default to what it ought to have been originally, and fixed
the constructor call accordingly.

Moved tabbed_browser from a local variable to a member of Crawler in order to
implement close().