In browser.js, pass in 'undefined' explicitly for an error code when there's no error code. Moved notify_load_end() upward in its function.
Fixed up a couple of 'yield' statements, removing parentheses needed as a
workaround for IDEA.

Wrote a Shim object to manage the replacement of 'Policy.processNode'. It's more
code than before, but it's already proven its worth in error checking during
development. It could also be extended to handle multiple replacements, and
transparent removal of a single shim, that is, if needed, since none of that
code is there yet.

Added a WeakMap 'current_nodes' to the crawler, to track nodes that it is
currently processing. Since the node_entry hook runs underneath the call stack
of the node hook, this allows us to associate filter activity with its node.

Closed some more resources in Crawler.close().

Added hook processing functions node_action and node_entry_action. This is an
internal port of the original functions processNode() and handleFilterHit().

Added onShutdown() for crawler.js.

Added an array 'observations' to the instruction type.