Added Logger.prototype.stack_trace(), to assist debugging.
Added a function intermediate to the crawler landing function. The new piece
dispatches the "real" landing function to the current thread, so it executes
after all code in the current thread finishes.

Split Tabbed_Browser.notify_load_end() into two pieces, one called at the end
of the load sequence, the other when the child closes. This leaves dispatch from
browser instances to our own code intact until we we're done with the tab,
which is at some point after it finishes loading in general (we want to monitor
script activity onLoad(), for example).