New Storage_class and stubs for derived classes replace the code in storage.js.
Changed storage initialization in Instruction.basic() to ignore its storage
argument and use a Display_Log storage. This is temporary. Removed its
formatting argument, but it will come back later.

Removed Instruction.null_storage, a stub.

Expanded on error handling in the crawler main loop body. It probably deserves
to be a library function, since it puts up a call stack on exception, and the
call stack formatting leaves something to be desired.

Initialize the Display_Log class in crawler_ui.js.