Changed the default tab to "Crawler". No longer using the sandbox.
Changed the instruction set to a class, instead of just a raw iterator. This
allows its metadata to be gathered neatly.

Added 'storage' argument to the Crawler class. This is where the argument really
belongs, since the Crawler represents a session, and the top-level output is
about a session, not simply the list of sites within it. Added 'toJSON' and
'output_encoding' along the same lines. Added code for a trial version of the
top-level output.

Made tracing in Format_Stream a little easier to manage.

Added code to support 'toJSON' replacement objects in the encode, as well as
'__encoding__' specifications.