Added number_of_tabs to the UI. Pass it to a new argument in the crawler. Used that value instead of the constant 1 to create the tabbed_browser member.
Added Tabbed_Browser.quiescent(), which indicates that the browser tab set has
no open tabs.

Added a loop in the crawler that waits for all pending tabs to close. It appears
immediately after the main instruction loop.

Added a couple of missing return statements in the node shim to return at the
right point when there's a processing exception.

Changed throw statements in the function to navigate from a window to a tab
instance to throw Error objects rather than raw strings.

Fixed a defect with merging the observation list.